% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{Apply a function to qualifying columns in a data frame, a different function otherwise}
\usage{
df_apply(.data, .f, .condition, .else = I, ...)
}
\arguments{
\item{.data}{Data frame to apply functions to}

\item{.f}{Function to apply to columns that fit .condition}

\item{.condition}{Condition to check to apply .f()}

\item{.else}{Function to apply to columns that do NOT fit .condition}

\item{...}{Additional arguments to be passed to .f()}
}
\value{
A Data frame with the columns changed as specified.
}
\description{
Apply a function to qualifying columns in a data frame, a different function otherwise
}
\examples{
n <- 7L
test_data <-
 data.frame(
   double     = rnorm(n, 100, 10),
   integer    = (1L:n) * (1L:n),
   character  = LETTERS[1L:n],
   factor     = factor(letters[1L:n]),
   logical    = rep(c(TRUE, FALSE), length.out = n)
 )
df_apply(test_data, round, is.numeric, I, digits = 2)
}
